# apps/backend/docs/external/player-api/openapi.v1.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/players/statistics/my": {
      "get": {
        "operationId": "StatisticsController_getMyStatistics_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get player statistics",
        "tags": [
          "Statistics"
        ]
      }
    },
...
    "/api/v1/messages/conversations": {
      "post": {
        "operationId": "MessagesController_createConversation_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateConversationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new conversation (player only)",
        "tags": [
          "Messages"
        ]
      },
      "get": {
        "operationId": "MessagesController_getConversations_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List user conversations",
        "tags": [
          "Messages"
        ]
      }
    },
    "/api/v1/messages/conversations/{id}": {
      "get": {
        "operationId": "MessagesController_getConversation_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get conversation details",
        "tags": [
          "Messages"
        ]
      }
    },
    "/api/v1/messages/conversations/{id}/messages": {
      "get": {
        "operationId": "MessagesController_getConversationMessages_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (default: 1)",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Messages per page (default: 50)",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get paginated messages for a conversation",
        "tags": [
          "Messages"
        ]
      },
      "post": {
        "operationId": "MessagesController_sendMessage_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Send a message in a conversation",
        "tags": [
          "Messages"
        ]
      }
    },
    "/api/v1/messages/conversations/{id}/read": {
      "post": {
        "operationId": "MessagesController_markAsRead_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mark a conversation as read",
        "tags": [
          "Messages"
        ]
      }
    },
    "/api/v1/messages/unread-count": {
      "get": {
        "operationId": "MessagesController_getUnreadCount_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get total unread messages count",
        "tags": [
          "Messages"
        ]
      }
    },
    "/api/v1/messages/admins": {
      "get": {
        "operationId": "MessagesController_getAdmins_v1",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search term to filter by username or email",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get list of admin users for conversation creation (player only)",
        "tags": [
          "Messages"
        ]
      }
    },
...
    "/api/v1/verification/request": {
      "post": {
        "operationId": "VerificationController_request_v1",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Request verification code",
        "tags": [
          "Verification"
        ]
      }
    },
    "/api/v1/verification/verify": {
      "post": {
        "operationId": "VerificationController_verify_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "code": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Verify with code",
        "tags": [
          "Verification"
        ]
      }
    },
    "/api/v1/players/reservations": {
      "post": {
        "operationId": "ReservationsController_create_v1",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Reserve a slot",
        "tags": [
          "Reservations"
        ]
      }
    },
    "/api/v1/players/reservations/{id}/confirm": {
      "post": {
        "operationId": "ReservationsController_confirm_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Confirm reservation",
        "tags": [
          "Reservations"
        ]
      }
    },
    "/api/v1/players/reservations/{id}": {
      "delete": {
        "operationId": "ReservationsController_cancel_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cancel reservation",
        "tags": [
          "Reservations"
        ]
      }
    },
    "/api/v1/players/reservations/my": {
      "get": {
        "operationId": "ReservationsController_getMyReservations_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get my reservations",
        "tags": [
          "Reservations"
        ]
      }
    },
...
  },
  "info": {
    "title": "BTV Player API",
    "description": "--",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "{baseUrl}",
      "variables": {
        "baseUrl": {
          "default": "http://localhost:3000"
        }
      }
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateConversationDto": {
        "type": "object",
        "properties": {
          "adminId": {
            "type": "number",
            "description": "ID of the admin to start conversation with"
          }
        },
        "required": [
          "adminId"
        ]
      },
...
      "SendMessageDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "description": "Message content",
            "maxLength": 10000
          }
        },
        "required": [
          "content"
        ]
      },
...
    }
  }
}